fix: Binding type mismatch and iOS 15 availability

+++ Sources/Echoelmusic/Views/VJLaserControlView.swift
@@ -1172,8 +1172,8 @@ struct SurfaceRowView: View {
 }
 
 struct KeystoneKnob: View {
-    @Binding var x: Float
-    @Binding var y: Float
+    @Binding var x: CGFloat
+    @Binding var y: CGFloat
 
     var body: some View {
         ZStack {
